+++ src/modules/CreateDrops/TicketDrop/CreateTicketDropForm.tsx
@@ -1,8 +1,10 @@
 import { useFormContext } from 'react-hook-form';
+import { Button, Divider, HStack } from '@chakra-ui/react';
 
 import { IconBox } from '@/common/components/IconBox';
 import { LinkIcon } from '@/common/components/Icons';
-import { MultiStepForm, StepItem } from '@/common/components/MultiStep/MultiStep';
+import { Step, StepItem } from '@/common/components/Step/Step';
+import { useSteps } from '@/common/hooks/useSteps';
 
 import { useDropFlowContext } from '../contexts/DropFlowContext';
 
@@ -40,12 +42,23 @@ const formSteps: StepItem[] = [
 export const CreateTicketDropForm = () => {
   const { onNext } = useDropFlowContext();
   const {
+    onNext: onNextStep,
+    onPrevious: onPreviousStep,
+    currentIndex,
+  } = useSteps({ maxSteps: formSteps.length });
+  const {
     setValue,
     handleSubmit,
     control,
     watch,
-    formState: { isDirty, isValid },
+    formState: { isDirty, isValid, errors },
   } = useFormContext();
+
+  console.log(errors);
+
+  const stepsDisplay = formSteps.map((step, index) => (
+    <Step key={step.name} index={index + 1} isActive={currentIndex === index} stepItem={step} />
+  ));
 
   const handleSubmitClick = () => {
     onNext();
@@ -57,7 +70,21 @@ export const CreateTicketDropForm = () => {
       maxW={{ base: '21.5rem', md: '36rem' }}
       mx="auto"
     >
-      <MultiStepForm steps={formSteps} />
+      <HStack flexWrap="nowrap" justifyContent="center" spacing="4">
+        {stepsDisplay}
+      </HStack>
+      {formSteps[currentIndex].component}
+      <Divider my={{ base: '6', md: '8' }} />
+      <HStack justifyContent="flex-end" spacing="auto">
+        {currentIndex > 0 && (
+          <Button variant="secondary" onClick={onPreviousStep}>
+            Go back
+          </Button>
+        )}
+        <Button variant="secondary" onClick={onNext}>
+          Skip this step
+        </Button>
+      </HStack>
     </IconBox>
   );
 };
